Move routing into its own package.
