+++ go/conversation_tmp/urls.py
@@ -20,5 +20,4 @@ urlpatterns = patterns(
     url(r'^incoming/(?P<campaign_key>\w+)/(?P<contact_key>\w+)/$',
         views.incoming_detail, name='incoming_detail'),
     url(r'^pricing/$', views.pricing, name='pricing'),
-    url(r'^routing/$', views.routing, name='routing'),
 )
