+++ go/conversation_tmp/views.py
@@ -221,29 +221,3 @@ def incoming_detail(request, campaign_key, contact_key):
 def pricing(request):
     return render(request, 'conversations/pricing.html', {
     })
-
-
-@login_required
-def routing(request):
-    # TODO: Better Go API client.
-
-    url = settings.GO_API_URL
-    auth = ('session_id', request.COOKIES['sessionid'])
-    req_data = {
-        "params": [request.user_api.user_account_key],
-        "jsonrpc": "2.0",
-        "method": "routing_table",
-        "id": None,
-    }
-    data = json.dumps(req_data)
-
-    r = requests.post(url, auth=auth, data=data)
-
-    model_data = {
-        'campaign_id': request.user_api.user_account_key,
-    }
-    model_data.update(r.json['result'])
-
-    return render(request, 'conversations/routing.html', {
-        'model_data': json.dumps(model_data),
-    })
