+++ go/urls.py
@@ -34,6 +34,10 @@ urlpatterns = patterns(
     url(r'^contacts/', include('go.contacts.urls', namespace='contacts')),
     url(r'^account/', include('go.account.urls', namespace='account')),
     url(r'^accounts/', include('registration.backends.default.urls')),
+
+    url(r'^routing/$', 'go.routing.views.routing', name='routing'),
+
+    # proxy API calls
     url(r'^api/', include('go.api.urls', namespace='api')),
 )
